move impl_expected_batch_read_hook macro
