+++ open-coroutine-core/src/lib.rs
@@ -95,7 +95,10 @@ pub mod net;
     clippy::not_unsafe_ptr_arg_deref,
     clippy::many_single_char_names,
     clippy::cast_sign_loss,
-    clippy::cast_possible_truncation
+    clippy::cast_possible_truncation,
+    clippy::cast_possible_wrap,
+    clippy::unnecessary_cast,
+    trivial_numeric_casts
 )]
 #[cfg(all(unix, feature = "syscall"))]
 pub mod syscall;
