+++ open-coroutine-core/src/syscall/facade.rs
@@ -6,7 +6,7 @@ use crate::syscall::raw::RawLinuxSyscall;
 #[cfg(target_os = "linux")]
 use crate::syscall::LinuxSyscall;
 use crate::syscall::UnixSyscall;
-use crate::{impl_expected_read_hook, impl_read_hook};
+use crate::{impl_expected_batch_read_hook, impl_expected_read_hook, impl_read_hook};
 #[cfg(target_os = "linux")]
 use libc::epoll_event;
 use libc::{
@@ -388,8 +388,10 @@ pub extern "C" fn readv(
     iov: *const iovec,
     iovcnt: c_int,
 ) -> ssize_t {
-    //todo
-    CHAIN.readv(fn_ptr, fd, iov, iovcnt)
+    unbreakable!(
+        impl_expected_batch_read_hook!(CHAIN, readv, fn_ptr, fd, iov, iovcnt,),
+        readv
+    )
 }
 
 #[must_use]
@@ -400,7 +402,10 @@ pub extern "C" fn preadv(
     iovcnt: c_int,
     offset: off_t,
 ) -> ssize_t {
-    CHAIN.preadv(fn_ptr, fd, iov, iovcnt, offset)
+    unbreakable!(
+        impl_expected_batch_read_hook!(CHAIN, preadv, fn_ptr, fd, iov, iovcnt, offset),
+        preadv
+    )
 }
 
 #[must_use]
@@ -410,7 +415,100 @@ pub extern "C" fn recvmsg(
     msg: *mut msghdr,
     flags: c_int,
 ) -> ssize_t {
-    CHAIN.recvmsg(fn_ptr, fd, msg, flags)
+    unbreakable!(
+        {
+            let blocking = is_blocking(fd);
+            if blocking {
+                set_non_blocking(fd);
+            }
+            let msghdr = unsafe { *msg };
+            let mut vec = std::collections::VecDeque::from(unsafe {
+                Vec::from_raw_parts(
+                    msghdr.msg_iov,
+                    msghdr.msg_iovlen as usize,
+                    msghdr.msg_iovlen as usize,
+                )
+            });
+            let mut length = 0;
+            let mut pices = std::collections::VecDeque::new();
+            for iovec in &vec {
+                length += iovec.iov_len;
+                pices.push_back(length);
+            }
+            let mut received = 0;
+            let mut r = 0;
+            while received < length {
+                // find from-index
+                let mut from_index = 0;
+                for (i, v) in pices.iter().enumerate() {
+                    if received < *v {
+                        from_index = i;
+                        break;
+                    }
+                }
+                // calculate offset
+                let current_received_offset = if from_index > 0 {
+                    received.saturating_sub(pices[from_index.saturating_sub(1)])
+                } else {
+                    received
+                };
+                // remove already received
+                for _ in 0..from_index {
+                    _ = vec.pop_front();
+                    _ = pices.pop_front();
+                }
+                // build syscall args
+                vec[0] = iovec {
+                    iov_base: (vec[0].iov_base as usize + current_received_offset) as *mut c_void,
+                    iov_len: vec[0].iov_len - current_received_offset,
+                };
+                cfg_if::cfg_if! {
+                    if #[cfg(any(
+                        target_os = "linux",
+                        target_os = "l4re",
+                        target_os = "android",
+                        target_os = "emscripten"
+                    ))] {
+                        let len = vec.len();
+                    } else {
+                        let len = c_int::try_from(vec.len()).unwrap();
+                    }
+                }
+                let mut new_msg = msghdr {
+                    msg_name: msghdr.msg_name,
+                    msg_namelen: msghdr.msg_namelen,
+                    msg_iov: vec.get_mut(0).unwrap(),
+                    msg_iovlen: len,
+                    msg_control: msghdr.msg_control,
+                    msg_controllen: msghdr.msg_controllen,
+                    msg_flags: msghdr.msg_flags,
+                };
+                r = CHAIN.recvmsg(fn_ptr, fd, &mut new_msg, flags);
+                if r != -1 {
+                    reset_errno();
+                    received += r as usize;
+                    if received >= length || r == 0 {
+                        r = received as ssize_t;
+                        break;
+                    }
+                }
+                let error_kind = std::io::Error::last_os_error().kind();
+                if error_kind == std::io::ErrorKind::WouldBlock {
+                    //wait read event
+                    if EventLoops::wait_read_event(fd, Some(Duration::from_millis(10))).is_err() {
+                        break;
+                    }
+                } else if error_kind != std::io::ErrorKind::Interrupted {
+                    break;
+                }
+            }
+            if blocking {
+                set_blocking(fd);
+            }
+            r
+        },
+        recvmsg
+    )
 }
 
 /// write
